Fix JstdConfigurationType factory default `getId`
com.intellij.diagnostic.PluginException: The default implementation of method 'getId' is deprecated, you need to override it in 'class com.google.jstestdriver.idea.execution.JstdConfigurationType$1'. The default implementation delegates to 'getName' which may be localized, but return value of this method must not depend on current localization. [Plugin: JSTestDriver Plugin]